+++ packages/create-pkg/src/index.ts
@@ -7,15 +7,18 @@ import fs from 'fs-extra';
 import path from 'path';
 import inquirer from 'inquirer';
 import { downloadMaterialTemplate, generateMaterial } from '@iceworks/generate-material';
+import getInfo from './langs/index.js';
 import { checkEmpty } from './checkEmpty.js';
-import { inquirPackageName } from './inquirPackageName.js';
+import { inquireProjectType } from './inquireProjectType.js';
+import { inquirePackageName } from './inquirePackageName.js';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 const cli = cac('@ice/create-pkg');
 
 (async () => {
-  cli.command('[...args]', 'Target dirname to generate', {
+  const info = await getInfo();
+  cli.command('[...args]', info.targetDir, {
     allowUnknownOptions: false,
     ignoreOptionDefaultValue: true,
   })
@@ -30,7 +33,7 @@ const cli = cac('@ice/create-pkg');
 
   const pkgPath = path.join(__dirname, '../package.json');
   const { version } = JSON.parse(fs.readFileSync(pkgPath, 'utf-8'));
-  consola.info('@ice/create-pkg version: ', version);
+  consola.info(info.version, version);
 
   cli.version(version);
 
@@ -42,6 +45,7 @@ const cli = cac('@ice/create-pkg');
   });
 
 async function create(dirPath: string, dirname: string): Promise<void> {
+  const info = await getInfo();
   await fs.ensureDir(dirPath);
   const empty = await checkEmpty(dirPath);
 
@@ -49,8 +53,7 @@ async function create(dirPath: string, dirname: string): Promise<void> {
     const { go } = await inquirer.prompt({
       type: 'confirm',
       name: 'go',
-      message:
-          'Files exist in the current directory already. Are you sure to continue ？',
+      message: info.dirExistFiles,
       default: false,
     });
     if (!go) process.exit(1);
@@ -58,9 +61,11 @@ async function create(dirPath: string, dirname: string): Promise<void> {
 
   const tempDir = path.join(dirPath, '.tmp');
 
-  await downloadMaterialTemplate(tempDir, '@ice/template-pkg-react');
+  const projectType = await inquireProjectType();
 
-  const npmName = await inquirPackageName();
+  await downloadMaterialTemplate(tempDir, `@ice/template-pkg-${projectType}`);
+
+  const npmName = await inquirePackageName();
 
   await generateMaterial({
     rootDir: dirPath,
@@ -74,7 +79,7 @@ async function create(dirPath: string, dirname: string): Promise<void> {
   await fs.remove(tempDir);
 
   console.log();
-  console.log('Initialize component successfully.');
+  console.log(info.initSuccess);
   console.log();
   console.log(`    cd ${dirname}`);
   console.log('    npm install');