+++ packages/pkg/src/helpers/normalizeRollupConfig.ts
@@ -5,12 +5,15 @@ import { isFile, findDefaultEntryFile } from '../utils.js';
 import commonjs from '@rollup/plugin-commonjs';
 import { nodeResolve } from '@rollup/plugin-node-resolve';
 import postcss from 'rollup-plugin-postcss';
+import PostcssPluginRpxToVw from 'postcss-plugin-rpx2vw';
 import alias from '@rollup/plugin-alias';
 import autoprefixer from 'autoprefixer';
 import swcPlugin from '../plugins/swc.js';
 import dtsPlugin from '../plugins/dts.js';
 import minify from '../plugins/minify.js';
 import babelPlugin from '../plugins/babel.js';
+import aliasPlugin from '../plugins/transform/alias.js';
+import rpx2vwPlugin from '../plugins/transform/rpx2vw.js';
 import { builtinNodeModules } from './builtinModules.js';
 import { TaskName } from '../types.js';
 
@@ -26,11 +29,13 @@ interface PkgJson {
 }
 
 const getRollupOutputs = ({
+  globals,
   userConfig,
   pkg,
   outputDir,
   isES2017,
 }: {
+  globals: Record<string, string>;
   userConfig: UserConfig;
   outputDir: string;
   pkg: PkgJson;
@@ -50,6 +55,7 @@ const getRollupOutputs = ({
     const commonOption = {
       name,
       format,
+      globals,
       sourcemap: sourceMaps,
     };
 
@@ -71,7 +77,7 @@ const getRollupOutputs = ({
   return outputs;
 };
 
-const getExternalsAndGlboals = (userConfig: UserConfig, pkg: PkgJson) => {
+const getExternalsAndGlboals = (userConfig: UserConfig, pkg: PkgJson): [(id?: string) => boolean, Record<string, string>] => {
   let externals: string[] = [];
   let globals: Record<string, string> = {};
 
@@ -117,7 +123,6 @@ export const normalizeRollupConfig = (
 ): [RollupPlugin[], RollupOptions] => {
   const { swcCompileOptions, type, outputDir, rollupPlugins, rollupOptions } = cfg;
   const { rootDir, userConfig, pkg } = ctx;
-
   const commonPlugins = [
     userConfig?.alias && alias({ entries: userConfig.alias }),
     userConfig?.babelPlugins?.length && babelPlugin(userConfig.babelPlugins),
@@ -125,7 +130,6 @@ export const normalizeRollupConfig = (
       extraSwcOptions: swcCompileOptions,
     }),
   ].filter(Boolean);
-
   let resolvedPlugins = rollupPlugins ?? [];
 
   if (type === 'transform') {
@@ -135,6 +139,10 @@ export const normalizeRollupConfig = (
       dtsPlugin(cfg.entry, userConfig?.generateTypesForJs),
       ...resolvedPlugins,
       ...commonPlugins,
+      aliasPlugin({
+        alias: userConfig?.alias,
+      }),
+      rpx2vwPlugin(),
     ];
 
     return [resolvedPlugins, rollupOptions];
@@ -145,7 +153,7 @@ export const normalizeRollupConfig = (
       ...resolvedPlugins,
       ...commonPlugins,
       postcss({
-        plugins: [autoprefixer()],
+        plugins: [autoprefixer(), PostcssPluginRpxToVw],
         extract: resolve(rootDir, outputDir, 'index.css'),
         autoModules: true,
         minimize: true,
@@ -171,8 +179,8 @@ export const normalizeRollupConfig = (
       {
         input: entry,
         external,
-        globals,
         output: getRollupOutputs({
+          globals,
           userConfig,
           pkg: pkg as PkgJson,
           outputDir: cfg.outputDir,