+++ packages/create-pkg/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0
+
+- [feat] support rax component/web library/node module project type
+- [feat] support language locale
+
 ## 1.0.1
 
 - [fix] remove `@@appworks/cli`.